refactor:Added Authorization policies to Complaints

+++ app/Http/Controllers/Complaint/ComplaintController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Complaint;
 
 use App\Http\Controllers\ApiController;
 use App\Models\Complaint;
+use App\Models\User;
 use App\Http\Requests\ComplaintStoreRequest;
 use App\Http\Requests\ComplaintUpdateRequest;
 use Illuminate\Support\Facades\Storage;
@@ -12,7 +13,7 @@ class ComplaintController extends ApiController
     public function __construct()
     {
         $this->middleware('auth:sanctum')->only([
-            'store','update','destroy'
+            'show','store','update','destroy'
         ]);
     }
     /**
@@ -33,6 +34,7 @@ class ComplaintController extends ApiController
      */
     public function store(ComplaintStoreRequest $request)
     {
+        $this->authorize('create',Complaint::class);
         $data=$request->only(['subject','body','place','district_id','priority']);
         $data['user_id']=auth()->user()->id;
         $data['status']='submitted';
@@ -52,6 +54,7 @@ class ComplaintController extends ApiController
      */
     public function show(Complaint $complaint)
     {
+        $this->authorize('view',$complaint);
         return $this->showModelAsResponse($complaint);
     }
 
@@ -64,6 +67,7 @@ class ComplaintController extends ApiController
      */
     public function update(ComplaintUpdateRequest $request, Complaint $complaint)
     {
+        $this->authorize('update',$complaint);
         $complaint->update($request->only([
             'subject','body','place','district_id','priority',
         ]));
@@ -87,6 +91,7 @@ class ComplaintController extends ApiController
      */
     public function destroy(Complaint $complaint)
     {
+        $this->authorize('delete',$complaint);
         $complaint->delete();
         return $this->showModelAsResponse($complaint);
     }
